Router de boards funcionan OK todos los endpoints.
