+++ NODE/src/controllers/board.controller.js
@@ -77,6 +77,62 @@ class BoardController {
           res.status(500).send({ status: "error", message: error.message });
         }
       };
+
+
+    
+      
+
+    addHBL2Board = async (req, res) => {
+      try {
+        const { name, hbl_id } = req.params; 
+    
+        if (!name || !hbl_id) {
+          return res.status(400).send({
+            status: 'error',
+            message: 'Missing required fields: name and hbl_id are both required.'
+          });
+        }
+    
+        const boardArray  = await this.service.getBoard({ name });
+
+        // Verifica si `getBoard` devuelve un array y accede al primer elemento
+        const board = Array.isArray(boardArray) ? boardArray[0] : boardArray;
+    
+        if (!board) {
+          return res.status(404).send({
+            status: 'error',
+            message: `Board with name "${name}" not found.`
+          });
+        }
+
+        console.log(`board: ${JSON.stringify(board)}`);
+        console.log(`board.devices: ${JSON.stringify(board.devices)}`);
+    
+        // Check for duplicate hbl_id before update (optional, depends on your logic)
+        const existingHBL = board.devices.find(device => device.hbl_id === hbl_id);
+        if (existingHBL) {
+          return res.status(409).send({
+            status: 'error',
+            message: `hbl_id "${hbl_id}" already exists in this board.`
+          });
+        }
+    
+        board.devices.push({ hbl_id }); // Add the new hbl_id to the devices array
+        await board.save();
+    
+        res.status(200).send({
+          status: 'success',
+          message: `hbl_id "${hbl_id}" added to board "${name}" successfully.`,
+        });
+      } catch (error) {
+        console.error(error);
+        res.status(500).send({
+          status: 'error',
+          message: 'Internal server error. Please try again later.',
+        });
+      }
+    };
+      
     
     updateBoard =  async (req, res)=>{
         const {uid} = req.params
