+++ NODE/src/enviroment/.env.development
@@ -2,3 +2,4 @@ PORT=8080
 MODO=desarrollo
 MONGO_URL='mongodb+srv://lqu:Jphlions135@cluster0.a4edk8z.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0'
 LEVEL_CONSOLE_LOG = 'debug'
+JWT_SECRET='Secret-key'
