+++ NODE/src/routers/board.routes.js
@@ -7,6 +7,7 @@ const {
     getBoards,
     getBoardsbyName,
     createBoard,
+    addHBL2Board,
     updateBoard,
     deleteBoard
 } = new BoardController()
@@ -14,6 +15,7 @@ const {
 router.get('/',        getBoards )
 router.get('/:name',        getBoardsbyName )
 router.post('/',       createBoard) 
+router.post('/:name/add/:hbl_id', addHBL2Board)
 router.put('/:uid',    updateBoard) 
 router.delete('/:uid', deleteBoard)
