+++ NODE/src/routers/register.routes.js
@@ -1,72 +1,14 @@
 const express = require('express');
 const Router = express.Router;  // Destructuring assignment for brevity
-const User = require('../dao/models/users.model.js');
-const bcrypt = require('bcrypt');
+const authController = require('../controllers/auth.controller')
 
-
+//const controller = new authController()
 
 const router = Router()
-router.post("/", signUp);
-router.get("/", showRegister);
-
-async function signUp(req,res){
-    const {name, email, password,role} = req.body;
-    if (!name || !email || !password ){
-        req.app.set('error',)
-        return res.status(403).send({ success : false, message : "All fields are required"})
-    }
+router.post("/", authController.signUp.bind(authController));
+router.get("/", authController.showRegister);
 
 
-    const existingUser = await User.findOne({email});
-
-    if (existingUser){
-        //console.log(existingUser)
-        res.status(400)/*.json({error : 'Email already registered'})*/
-        //req.flash('error', `Email already registered`)
-        req.app.set('error', `Email already registered`)
-        return res.redirect('/api/register')
-    }
-    
-    let hashedPassword = await hashPassword(password);
-    
-    
-    try {
-        const newUser = new User({
-            name,
-            email,
-            password:hashedPassword,
-            role
-          })    
-        await newUser.save()
-        res.status(200)/*.json({message: `user registered successfully`})*/
-        //req.flash('success', `user registered successfully`)
-        req.app.set('success',`user registered successfully`)
-        return res.redirect('/api/login')
-    } catch (error) {
-        res.status(500)/*.json({"error" : error.message})*/
-        //req.flash('error', `${error.message}`)
-        req.app.set('error',`${error.message}`)
-        return res.redirect('/api/login')
-    }
-}
- async function hashPassword(password){
-    
-    try {
-        return await bcrypt.hash(password,10)
-        
-    } catch (error) {
-        return res.status(500).json({
-            success : false,
-            message : `Hashing password error` + error.message
-        })
-    }
-}
-
-function showRegister(req,res){
-
-    
-    return res.render('register')
-}
