+++ NODE/src/views/dashboard.hbs
